requirements.txt
streamlit
tensorflow-cpu
pillow
numpy
protobuf<4.0
EOF